* fix-ie-on-tests:
  Add browser matrix
  Fix, ie friendly message test failure
  Make it an explicit 404
  fix issue with stubs not being released after test
  Dynamically generate blade.yml
  Remove hacky extension, prefer to use the dummy application
  Remove superflous override.
  The blade helper needs to be included on startup.
  Fix ie11 can't have semicolons inside parenthesis
  Refactor use Relax.warn
  IE doesn't support payload keys
  Fix session isn't always there.
  iNeed to reset....Updating blade with the latest qunit broke sinon-qunit. Removing and just do it manually
  Fix IE relative urls in links come up blank in some attributes
  Move fixtures to rails dummy app

+++ lib/bath/handler.rb
@@ -31,7 +31,7 @@ module Bath
             json.path relax_filter
           end
 
-          if defined?(session)
+          if defined?(session) && session
             session.delete(:relax_filter)
           end
